requirements.txt
# ============================================================================
# SPA VIP - RENDER DEPLOYMENT REQUIREMENTS
# Optimized for production deployment on Render.com
# ============================================================================

# ============================================================================
# 🗄️ CORE INFRASTRUCTURE & DATABASE
# ============================================================================
supabase==2.17.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7

# ============================================================================
# 🕷️ WEB CRAWLING & DATA EXTRACTION
# ============================================================================
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1
lxml==4.9.3
python-dateutil==2.8.2

# ============================================================================
# 🤖 AI/ML MODELS & DEEP LEARNING (CPU-OPTIMIZED)
# ============================================================================
torch==2.0.1+cpu
transformers==4.35.2
tensorflow-cpu==2.13.0
sentencepiece==0.1.99
accelerate==0.24.1

# ============================================================================
# 📊 DATA PROCESSING & ANALYSIS
# ============================================================================
pandas==2.1.3
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# ============================================================================
# 📈 VISUALIZATION (LIGHTWEIGHT)
# ============================================================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# ============================================================================
# 🛠️ ESSENTIAL UTILITIES
# ============================================================================
tqdm==4.66.1
Pillow==10.1.0
typing-extensions==4.8.0
colorama==0.4.6
python-json-logger==2.0.7

# ============================================================================
# 🔧 PERFORMANCE & OPTIMIZATION
# ============================================================================
numexpr==2.8.8
bottleneck==1.3.7

# ============================================================================
# 🌐 WEB FRAMEWORK (FOR API ENDPOINTS)
# ============================================================================
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# ============================================================================
# � ADDITIONAL UTILITIES
# ============================================================================
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1

# ============================================================================
# 🔐 SECURITY & VALIDATION
# ============================================================================
pydantic==2.5.0
python-jose==3.3.0

# ============================================================================
# 📊 MONITORING & LOGGING
# ============================================================================
structlog==23.2.0

# ============================================================================
# 🎯 RENDER-SPECIFIC OPTIMIZATIONS:
# - CPU-only versions for faster deployment
# - Fixed versions for reproducible builds
# - Lightweight alternatives where possible
# - Web framework ready for API deployment
# ============================================================================
